Refactored file structure to be more modular
