+++ backend/object/models/__init__.py
@@ -1,0 +1,2 @@
+from .base_object import Object
+from .cabinet import Cabinet
